Move SpectrumEvaluator tests
