+++ gammapy/spectrum/tests/test_utils.py
@@ -1,17 +1,9 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-import pytest
-import numpy as np
 from numpy.testing import assert_allclose
 import astropy.units as u
 from astropy.units import Quantity
-from gammapy.irf import EffectiveAreaTable, EnergyDispersion
-from gammapy.modeling.models import (
-    ExponentialCutoffPowerLaw,
-    PowerLaw,
-    PowerLaw2,
-    TableModel,
-)
-from gammapy.spectrum import SpectrumEvaluator, integrate_spectrum
+from gammapy.modeling.models import ExponentialCutoffPowerLaw, PowerLaw
+from gammapy.spectrum import integrate_spectrum
 from gammapy.utils.testing import assert_quantity_allclose, requires_dependency
 
 
@@ -49,54 +41,3 @@ def test_integrate_spectrum_ecpl():
 
     assert res.unit == "cm-2 s-1"
     assert_allclose(res.value, [5.95657824e-13, 9.27830251e-14], rtol=1e-5)
-
-
-def get_test_cases():
-    e_true = Quantity(np.logspace(-1, 2, 120), "TeV")
-    e_reco = Quantity(np.logspace(-1, 2, 100), "TeV")
-    return [
-        dict(model=PowerLaw(amplitude="1e2 TeV-1"), e_true=e_true, npred=999),
-        dict(
-            model=PowerLaw2(amplitude="1", emin="0.1 TeV", emax="100 TeV"),
-            e_true=e_true,
-            npred=1,
-        ),
-        dict(
-            model=PowerLaw(amplitude="1e-11 TeV-1 cm-2 s-1"),
-            aeff=EffectiveAreaTable.from_parametrization(e_true),
-            livetime="10 h",
-            npred=1448.05960,
-        ),
-        dict(
-            model=PowerLaw(reference="1 GeV", amplitude="1e-11 GeV-1 cm-2 s-1"),
-            aeff=EffectiveAreaTable.from_parametrization(e_true),
-            livetime="30 h",
-            npred=4.34417881,
-        ),
-        dict(
-            model=PowerLaw(amplitude="1e-11 TeV-1 cm-2 s-1"),
-            aeff=EffectiveAreaTable.from_parametrization(e_true),
-            edisp=EnergyDispersion.from_gauss(
-                e_reco=e_reco, e_true=e_true, bias=0, sigma=0.2
-            ),
-            livetime="10 h",
-            npred=1437.450076,
-        ),
-        dict(
-            model=TableModel(
-                energy=[0.1, 0.2, 0.3, 0.4] * u.TeV,
-                values=[4.0, 3.0, 1.0, 0.1] * u.Unit("TeV-1"),
-            ),
-            e_true=[0.1, 0.2, 0.3, 0.4] * u.TeV,
-            npred=0.554513062,
-        ),
-    ]
-
-
-@pytest.mark.parametrize("case", get_test_cases())
-def test_counts_predictor(case):
-    opts = case.copy()
-    del opts["npred"]
-    predictor = SpectrumEvaluator(**opts)
-    actual = predictor.compute_npred().total_counts
-    assert_allclose(actual, case["npred"])
